fixing a bug connecting to a non prysm client and moving the prysm api call to the prysm beacon client
