+++ validator/client/beacon-api/BUILD.bazel
@@ -106,6 +106,7 @@ go_test(
         "propose_beacon_block_phase0_test.go",
         "propose_beacon_block_test.go",
         "propose_exit_test.go",
+        "prysm_beacon_chain_client_test.go",
         "registration_test.go",
         "state_validators_test.go",
         "status_test.go",
@@ -116,7 +117,6 @@ go_test(
         "subscribe_committee_subnets_test.go",
         "sync_committee_selections_test.go",
         "sync_committee_test.go",
-        "validator_count_test.go",
         "wait_for_chain_start_test.go",
     ],
     embed = [":go_default_library"],
